Cut clean-build time 4x by splitting the giant top-level Program.cs
Roslyn's interprocedural taint analyzers (CA3001-CA3012, CA5390, CA5403,
CA2000, enabled by AnalysisLevel=latest-all) scale super-linearly with a
single method body's size. GameBot.Service's 1,800-line top-level
Program.cs - where every endpoint lambda and post-Run() helper is part of
one implicit Main - cost ~306s of the ~354s total analyzer time, making
the clean solution build take ~6 minutes (~10 on CI).

Move the code verbatim into normal-sized members, no behavior change:
- DI registrations + URL binding -> GameBotServiceSetup.cs
- /sequences endpoints + helpers -> Endpoints/SequencesEndpoints.cs
- /versioning + /installer endpoints -> Endpoints/VersioningEndpoints.cs
- legacy route guards -> Endpoints/LegacyGuardEndpoints.cs
- Swagger document filter -> Swagger/ConditionalFlowSchemaDocumentFilter.cs

Also pin Microsoft.CodeAnalysis.NetAnalyzers to 10.0.301 (was floating
*, which silently changes rule behavior and defeats the CI NuGet cache
key), and point the three installer guard tests that grep Program.cs
source at the endpoints' new home.

Clean rebuild: 356s -> 89s with -warnaserror still on and 0 warnings;
all 941 tests pass. Analyzer rule coverage is unchanged.
